# Cache Factory._find_provider_class module look-ups, make Faker construction 20✕