--- a/faker/factory.py
+++ b/faker/factory.py
@@ -1,3 +1,4 @@
+import functools
 import locale as pylocale
 import logging
 import sys
@@ -64,6 +65,7 @@ class Factory:
         return faker
 
     @classmethod
+    @functools.cache
     def _find_provider_class(
         cls,
         provider_path: str,